+++ mypy/dmypy/client.py
@@ -46,7 +46,7 @@ parser.add_argument(
     "-V",
     "--version",
     action="version",
-    version=f"Basedmypy Daemon {__based_version__}\nBased on %(prog)s {__version__}",
+    version=f"Basedmypy-Daemon {__based_version__}\nBased on %(prog)s {__version__}",
     help="Show program's version number and exit",
 )
 subparsers = parser.add_subparsers()