+++ mypy/test/testcmdline.py
@@ -69,6 +69,10 @@ def test_python_cmdline(testcase: DataDrivenTestCase, step: int) -> None:
     if not based:
         args.append("--no-strict")
         args.append("--no-default-return")
+    if "--pretty" not in args:
+        args.append("--no-pretty")
+    if "--show-error-code-links" not in args and "--ide" not in args:
+        args.append("--hide-error-code-links")
     if "--error-summary" not in args:
         args.append("--no-error-summary")
     if "--show-error-codes" not in args and not based:
@@ -85,7 +89,7 @@ def test_python_cmdline(testcase: DataDrivenTestCase, step: int) -> None:
     env.pop("COLUMNS", None)
     extra_path = os.path.join(os.path.abspath(test_temp_dir), "pypath")
     env["PYTHONPATH"] = PREFIX
-    env["__MYPY_UNDER_TEST__"] = "1"
+    env["__MYPY_UNDER_TEST__"] = "1" if based else "2"
     if os.path.isdir(extra_path):
         env["PYTHONPATH"] += os.pathsep + extra_path
     cwd = os.path.join(test_temp_dir, custom_cwd or "")