+++ mypy/test/testerrorstream.py
@@ -29,6 +29,8 @@ def test_error_stream(testcase: DataDrivenTestCase) -> None:
     options = Options()
     options.show_traceback = True
     options.hide_error_codes = True
+    options.show_error_context = False
+    options.pretty = False
 
     logged_messages: list[str] = []