+++ mypy/test/testpep561.py
@@ -126,7 +126,15 @@ def test_pep561(testcase: DataDrivenTestCase) -> None:
                     f.write(f"{s}\n")
             cmd_line.append(program)
 
-        cmd_line.extend(["--no-error-summary", "--hide-error-codes", "--no-strict"])
+        cmd_line.extend(
+            [
+                "--no-error-summary",
+                "--hide-error-codes",
+                "--no-strict",
+                "--no-pretty",
+                "--hide-error-context",
+            ]
+        )
         if python_executable != sys.executable:
             cmd_line.append(f"--python-executable={python_executable}")